refactor: 패키지 구분을 위한 클래스 이동
