+++ src/main/java/lotto/domain/Money.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Money {
     public static final int LOTTO_PRICE = 1000;
