+++ src/test/java/lotto/LottoNumberGeneratorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
